Upper Triangular Matrix
The population of North Korea and South Korea are represented in a square matrix, with the upper triangular elements(including the diagonal) representing North Korea and lower Triangular elements(excluding the diagonal) representing South Korea. Kim Jong-un  asks you to inform him the current population figures of his country. Note: The elements are in a scale of 1000. 1 means 1000, 2 means 2000 and so on.
 

Input Description:
Dimensions of the matrix followed by the elements of the matrix

Output Description:
Sum of upper triangular elements